Loading Fix
"Page Icons" continue to show up as null and cause errors on the pages version of the site which isn't reflected on the locally hosted version of the site.

I think the likely culprit is that the async calls we make to load pageicons is fast enough locally to not cause any issues but on the hosted version we call the functions that rely on pageicons before the async function finishes.

Attempting to fix by setting the page to load only after async functions finish, this *may* cause some additional loadtime that will look ugly to the end user.  If it becomes an issue we should do another pass on the css