Local Saves Update
Adds two new buttosn in the workbar that allow for downloading the current profile as a local JSON save file and uploading said save files to the sheet.

This isn't *quite* a replacement for being able to tab between multiple profiles but it's a nice band-aid that lets us sidestep the issue of storage while also allowing for local backups in the case of a bug that would require clearing local storage.